Adapt VPA metrics to new solution CustomResourceState
This commit adds the necessary configuration changes to readd the VPA
metrics using the new strategy CustomResourceState from the
kube-state-metrics. Such configuration is created programmatically thanks
to the the kube-state-metrics dependency downloaded in the previous
commit. This commit also runs `go mod tidy` in order to make this
dependency direct.

The new CustomResourceState configuration is stored in a ConfigMap and
mounted into the kube-state-metrics deployment using the
`--custom-resource-state-config-file`. There is an alternative command
flag `--custom-resource-state-config` to pass such configuration inline
but the resulting configuration is very lengthy, which would make the
kube-state-metrics deployment document not so readable.

We tried to preserve the old metric labels and naming as much as
possible. However, there are few changes:

1. We can't have one common metric for several resources (e.g., CPU
  and memory) and distinguish between them via labels. Instead, we
  need to migrate to one specific metric per resource. For instance,
  previously we would have a `target` metric with labels `resource=cpu`
  or `resource=memory` to indicate the value corresponding to CPU or
  memory. After this change two metrics are added: `target_cpu` and
  `target_memory`. Nevertheless, we consider this as an improvement
  because values in the new metrics are not mixing units as before
  (cores were used for the CPU and bytes for the memory).
2. Using `CustomResourceState` prefixes the metric name with
   `kube_customresource` by default. Therefore, the old prefix
   `kube_verticalpodautoscaler` becomes now
   `kube_customresource_verticalpodautoscaler`. Such default prefix
   can be overwritten to match the one we want, but we choose not to
   so that the new name reflects better that the metric is generated
   by the `CustomResourceState`.
3. The property `nilIsZero` is set to true for the recommendation
   metrics to set the value to zero if the recommendation path does not
   exist in the VPA spec file (i.e., it doesn't have a recommendation).
   In the past, the time series would simply not exist. Since a
   recommendation to 0 does not make sense, this is a way, e.g., in the
   dashboard, to know there is no recommendation at all, rather than
   relying on not having data, which could also be the case if there is
   an issue and metrics are missing.
4. Metrics containing `container_policies` in their name have been
   renamed to `containerpolicies` to respect the containerPolicies key
   in the spec file (similar to resourcePolicy or
   containerRecommendations). This way, the metric format is unified to
   use underscores to separate keys in the path.
5. The `CustomResourceState` documentation does not mention that
   kube-state-metrics needs `get` permissions on RBAC-enabled clusters,
   but only `list` and `watch`. Therefore, the old `get` permissions are
   removed.

A unit test is added to assert the generated configuration matches the
one originally generated for testing. The test expectation is saved into
a yaml file and committed to git. The test saves the generated
configuration in a different file in the temporary folder (but does not
commit it). The benefit of this approach is that it helps comparing the
actual and expected configuration, no matter how large the files are: we
have two separate files that we can compare using any tool.

The `expectedCustomResourceStateConfig()` utility returns the expected
CustomResourceState config and also asserts that the actual value is the
same. This function is to be used to load the test expectation during
the test setup. The assertion is performed inside this function to allow
to give more human readable errors when the long config document actually
differs. When the assertion fails, a custom message is shown. It tries to
mimic the wording from usual ginkgo test runs but adds a hint in the end
on how to use the diff command to see the diff. This of course will only
work on UNIX systems. Developers can review this diff and react
accordingly e.g., check if there is a bug in the code or if indeed the
expectation has to be changed.

Finally, this commit also runs `go mod tidy` to make the new
kube-state-metrics dependency direct.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>