Change namespace used in metric for control-plane shoot pods
The kube-system namespace was previously used. This was wrong in two ways.
First, those pods are not deployed in the kube-system of the shoot, but
in a specific namespace in the seed. Second, there is a blackbox
exporter in both the shoot's namespace in the seed and in the actual
kube-system namespace in the shoot. This led to name conflicts and
possibly unexpected query results. This is changed to shoot-control-plane.