Change namespace used in metric for control-plane shoot pods
The kube-system namespace was previously used. This was wrong in two ways.
First, those pods are not deployed in the kube-system of the shoot, but
in a specific namespace in the seed. Second, there is a blackbox
exporter in both the shoot's control-plane namespace in the seed and in
the actual kube-system namespace in the shoot. This led to naming conflicts and
possibly unexpected query results, e.g. in the VPA dashboard when checking the
VPA recommendations for the blackbox_exporter pods. The namespace is changed to
shoot-control-plane, which properly describe what it is.

Note that the control plane namespace is technically called
`shoot--<project>--<shoot name>`, but instead of leaking this implementation
detail to the monitoring stack, we use the more user-friendly name
`shoot-control-plane`.