Fix VPA recommendations dashboards for the Prometheus custom resource
Since the introduction of the Prometheus operator as part of GEP-19, VPA
targets the Prometheus custom resource. Prometheus pods are ever since
prefixed with prometheus-, which breaks the VPA dashboard because the
targetName is not enough anymore for the pattern matching. We also need
the targetKind to match the pod name. We use flag `?i` to match
case-insenstively, given that "kinds" in kubernetes are titled but pods
are not.

Example: kind is Prometheus but pod name is prometheus-shoot-0.