WIP Separate scrape configs
WIP the scrape config names
- garden-kube-state-metrics-runtime
- seed-kube-state-metrics-seed

contain both a prefix (garden/seed) and a suffix (runtime/seed) that serve a
similar purpose but are slightly different. Could this be aligned somehow? Maybe
we could use only the prefix for these resources to retain the previous
behaviour of having a prefix, but change the prefix garden->runtime to align
with the terminology. We anticipate having both a garden and a runtime
kube-state-metrics instance in the garden namespace, so we will have both
prefixes: garden and runtime in the future.