Indent tool installation commands
Each installed tool starts with a new echo entry, followed by the
required commands. These are over-indented with respect to the echo
entry so that the echo looks like a title and then then sequence of
commands. A VERSION variable might be created to shorten some lines,
similar as it's done on the version from Istvan's branch.