Remove previous image only if it exists
Previous code could not work when the cell image did not exist at all.
In such case, the docker rmi command returns false and the next command
is not evaluated because of the &&.

The command to remove the old cell image is now also embedded in the
docker exec command block that builds the image.