Handle the case when the cell image does not exist yet
Inspired by e8dc3c3 Remove previous image only if it exists

Note that the image should be deleted before the image is started in the
next pane: if the image were deleted in the last command of this pane,
we would need some other synchronization mechanisms. By deleting the image
in the last but one command, it will be executed before the first command
of the next pane.