Fix Merge the garden and the KinD kube configs
The virtual garden's kubeconfig is watched on the 'token refresh' pane.
When the secret changes, the kubeconfig is written to
'example/operator/virtual-garden/kubeconfig'.

To prevent inconsistent reads, the kubeconfig is first written to
example/operator/virtual-garden/kubeconfig.wip
and then the '.wip' file is renamed to
example/operator/virtual-garden/kubeconfig.
The rename is atomic, so it should prevent inconsistent reads.

In Merge the garden and the KinD kube configs,
inotifywait is used to wait for changes to this file efficiently.
The kubeconfig file is not written to, but moved to, so the
previously used event "close_write" was incorrect and that filter condition
was never met.

This led to the symptom that the merged kubeconfig was never updated,
and the token timed out after 12h.

Note that in the cell, the example/operator/virtual-garden/kubeconfig
was properly updated, only the merged ~/.kube/config was affected.

After this fix, even the merged kubeconfig is going to be properly updated.

To debug the inotify events in the cell:

[~/go/src/github.com/gardener/gardener (⎈|kind-gardener-operator-local:N/A)] ((v1.103.0)) [10:32]# docker exec -it jail docker exec -it cell bash
[~/go/src/github.com/gardener/gardener (⎈|garden:garden-local)] ((v1.103.0))
$ inotifywait -m /home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden
Setting up watches.
Watches established.

/home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden/ CREATE kubeconfig.wip
/home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden/ OPEN kubeconfig.wip
/home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden/ MODIFY kubeconfig.wip
/home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden/ MODIFY kubeconfig.wip
/home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden/ CLOSE_WRITE,CLOSE kubeconfig.wip
/home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden/ MOVED_FROM kubeconfig.wip
/home/demo/go/src/github.com/gardener/gardener/example/operator/virtual-garden/ MOVED_TO kubeconfig

So there is a moved_to event, but no close write event.

To test it, edit the gardener secret in the garden namespace:

k edit secret -n garden gardener and observe that the merged kubeconfig is updated.