Adapt the kube readonly panes
This pane should be revisited as the original kubeconfig created by
kind is valid inside the cell. Therefore, only the token is probably
required to be added. The SAN is only necessary to test this new demo
kubeconfig with token as root outside docker.

This pane used to prepare a hosts.entry with the virtual and runtime
cluster IPs. This file was then passed on to the cell but this is no
longer necessary: the cell is in the kind network so it can resolve the
'gardener-operator-local-control-plane' name on its own.