Store the objects in the objectsToGarbageCollect map
Previously, the map was used as a set for the object IDs: the object ID was
stored as a key in the map, and the value was an empty struct.

Now, we store the objects in the map as a value.

Later, we'll need the resource version and the UID fields to create the
`DeleteOptions` object with the preconditions. By storing the objects in the
map, we can avoid querying the API server for these objects again.