Use the object from the map for the deletion
Now that we store the objects in the `objectsToGarbageCollect` map, we can use
them when deleting these objects.

Previously a corev1.Secret or corev1.ConfigMap object was created, and was
passed to the `Delete` method of the client. Now we use the
PartialObjectMetadata object from the map to delete the object.