Detect conflicts in the Delete() function
A plain delete request can not lead to a 409 Conflict because the UID and the
resourceVersion are not passed in the underlying HTTP delete request to the API
server.

The Kubernetes API allows to pass a `DeleteOptions` object in the request body
with the `preconditions.{resourceVersion,uid}` fields to assert that the
resource to be deleted has not been changed in the meantime.

See [kubernetes-api#delete-secret](https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.31/#delete-secret-v1-core)

This change adds the `DeleteOptions` to the `Delete` call in the Garbage
Collector reconciler to detect conflicts when deleting resources.