Mark objects as unused before deleting them
The Garbage Collector sets the new `gc.gardener.cloud/unusedAt` label to mark
resources as unused, and deletes them only if they already have this label.

Previously, the Garbage Collector deleted the unused resources right away, but
this can lead to race conditions.

Now, the Garbage Collector first marks the resources as unused by adding the
`gc.gardener.cloud/unusedAt` label with the current timestamp, if the resource
does not have the `gc.gardener.cloud/unusedAt` label yet, and it does not delete
the resource right away.

The Garbage Collector also sets the `gc.gardener.cloud/used` label to `false`.
When a client reuses the resource, it can set the `gc.gardener.cloud/used` label
back to `true`, thereby mutating the resource. This allows the Garbage Collector
to detect a race condition when deleting the resource.

The object is a PartialObjectMetadata object. The client does not support the
Update method for PartialObjectMetadata objects. Therefore, the Patch method is
used with the optimistic lock strategy. The optimistic lock strategy ensures
that the resource version of the object is sent with the patch request.

https://github.com/kubernetes-sigs/controller-runtime/blob/v0.19.2/pkg/client/client.go#L289-L290
https://github.com/kubernetes-sigs/controller-runtime/blob/v0.19.2/pkg/client/patch.go#L104-L115