Requeue soon if some resources were marked as unused
Instead of deleting unused resources right away, first they are marked with the
`gc.gardener.cloud/unusedAt` label as to be deleted, and then a few minutes
later, if they are still not referenced, they can be finally deleted.

The Garbage Collector is executed infrequently, typically on startup and every
12 hours.

When the Garbage Collector marks some resources as unused, we should requeue the
reconciliation sooner than the typical interval of 12 hours, e.g. in a few
minutes, so that the deletion of the resources can be finalized in a timely
manner.

The resources are marked as unused in a separate goroutine, so a channel is used
to relay this information to the main goroutine of the Garbage Collector,
similar to the existing results channel.

The existing results channel collects the errors and the new requeueNeeded
channels collects true values if the Garbage Collector should be requeued soon.

To avoid blocking when these channels are read in the main goroutine, both
channels are read concurrently in a loop until both channels are closed.

The previously used for loop with a range over the results channel:

for err := range results {
	...
}

has to be replaced with a slightly more complicated construct, a combination of
a loop and a select statement, to read from both channels until they are closed.

var err error
resultsOK := true
requeueNeededOK := true
for resultsOK || requeueNeededOK {
	select {
	case err, resultsOK = <-results:
		... // append err to a list
	case _, requeueNeededOK = <-requeueNeeded:
		...
	}
}