Skip objects that were marked as unused only recently
This is similar to the

  if obj.CreationTimestamp.Add(*r.MinimumObjectLifetime).UTC().After(r.Clock.Now().UTC()) {
  // Do not consider recently created objects for garbage collection.
  continue
  }

check in the first part of the file.

What is the rationale behind this?

Kubernetes does not offer mechanisms to have a consistent view on multiple
resources at a given point in time.

If an actor changes multiple resources, e.g. it creates a managed resource and a
secret, we can not rely on the ordering of these operations. So the Garbage
Collector might observe the newly created secret but might not yet see the
managed resource.

Eventually, the Garbage Collector will observe both the managed resource and the
secret.

To avoid that the Garbage Collector considers a secret to be unused and delete
it due to the delay of observing the referring managed resource, recently
created secrets are not considered for garbage collection. The rationale is that
if a secret was created more than the minimum object lifetime (10 minutes) ago,
it is expected that the Garbage Collector has observed all the resources that
reference the secret.

Similarly, if a secret was marked for deletion only recently, it is not going to
be considered for garbage collection. If the secret was marked for deletion more
than the minimum object lifetime ago, it is expected that the Garbage Collector
has observed all the resources that reference the secret and hence an
unreferenced secret can be safely deleted.