Remove the unusedAt label from used objects
A silent race condition between adding the unusedAt label to an object and
referencing the object is still possible.

Previously, the Garbage Collector simply deleted the object, which left the
system in an inconsistent state.

Now, the Garbage Collector only adds the unusedAt label to the object, but does
not delete it right away. The deletion takes place a few minutes later.

So if the object is referenced again in the meantime, the Garbage Collector will
remove the unusedAt label and set the used label back to true.