Add documentation for the Cooperative Executor
This PR will later use the *Cooperative Executor* struct of the 'race' test
utility package to check for race conditions in the Garbage Collector
implementation.

This commit adds a documentation that introduces the problem of race conditions
based on a simple example and then shows how the Cooperative Executor (that will
be implemented in the next commit) can help to detect race conditions in unit
tests.