TODO Add a test to show the race condition
TODO: list the managed resources
TODO: how to simulate that the secret is used, but the managed resource is not yet updated.

This commit adds a unit test that shows that the Garbage Collector
of the Gardener Resource Manager has a race condition.

It uses the Cooperative Executor that was introduced in the previous commits.

The Cooperative Executor is used to run the Garbage Collector and an actor that
reconciles a managed resource and a secret, with the managed resource
referencing the secret. The Garbage Collector and this actor are executed
concurrently. The Kubernetes client API calls are interleaved in all possible
ways.

In the initial state of the system there is one unreferenced secret. The unit
test shows that there is an ordering of the API calls where the Garbage
Collector deletes the secret that is "reused" by the actor. Neither the Garbage
Collector, nor the actor detects an error, and hence the system is left in an
inconsistent state indefinitely.

Why "reused"? Immutable secrets have content based names. If the actor
reconciles a secret with the same content, it will reuse already existing secret
with the same name.

The purpose of this PR is to prevent the "silent" deletion of such secrets by
the Garbage Collector. The fix is to detect conflicts between the Garbage
Collector's delete method call and the client "reusing" the secret.

The unit test shows the 3 orderings of the relevant API calls:

Path 1:
- GC/0 List secrets
- GC/1 Delete unused secret
- Actor/0 Get secret --> not found
- Actor/1 Create secret

In this path, the secret is deleted by the Garbage Collector and recreated by
the actor.

Path 2:
- Actor/0 Get secret --> the immutable secret exists
                         no need to create or update
- GC/0 List secrets
- GC/1 Delete unused secret

TODO

Path 3 is similar to Path 2:
- GC/0 List secrets
- Actor/0 Get secret --> the immutable secret exists
                         no need to create or update
- GC/1 Delete unused secret

The GC deletes the secret, although it is referenced in the meantime.