Define new labels: resources.gardener.cloud/{used,unused_at}
Clients are expected to set the label
- `resources.gardener.cloud/used` to `true`
on the immutable secrets or configmaps that are referenced by other resources
and that are managed by the Garbage Collector.

The Garbage Collector will set the labels
- `resources.gardener.cloud/used` to `false`, and
- `resources.gardener.cloud/unused_at` to the current time
when it marks a resource as "to be deleted".

Hence, when a client reuses a resource that was marked to
be deleted, the client will mutate the resource:
- `resources.gardener.cloud/used` `false` --> `true`
and that can be used to detect a conflict when the Garbage Collector
deletes the resource.