Set resources.gardener.cloud/used to true in MakeUnique
The utils.kubernetes/MakeUnique function is the only place in the
gardener/gardener code base that sets the

  resources.gardener.cloud/garbage-collectable-reference

label on secrets and configmaps when it makes them immutable.

After setting the

  resources.gardener.cloud/garbage-collectable-reference

label, the GRM's Garbage Collector will be responsible for deleting these
secrets and configmaps when they are not referenced by any more.

This PR introduces a new label that clients are expected to set:

  resources.gardener.cloud/used = true

Hence, the utils.kubernetes/MakeUnique functions is a convenient, place to make
sure that all the clients that create immutable secrets and configmaps set the
new label as expected by this PR.

Note that the approach of this PR is backwards compatible:

Some clients (e.g. extensions) will not set the label until this change
propagates to them as well. In that case the previous behavior (a small chance
for race conditions) is preserved.

So this change is backwards compatible, it does not require that all the clients
use the new label.