Configure a controller runtime logger in the unit test
While debugging the unit test and spending more than 30 seconds in breakpoints I
noticed that the controller runtime's logger emits a warning:

> [controller-runtime] log.SetLogger(...) was never called; logs will not be displayed.
> Detected at:%s%s

https://github.com/kubernetes-sigs/controller-runtime/blob/v0.19.2/pkg/log/log.go#L65

By configuring a logger for the controller runtime in the unit test, this
warning can be avoided and the logs that e.g. the Garbage Collector emits are
shown in the standard out.