Allow the Gardenlet to see the ManagedSeed object of its own cluster (part 2)
PR https://github.com/gardener/gardener/pull/11965 attempted to
fix an issue that was introduced in Gardener v1.117.0 with
https://github.com/gardener/gardener/pull/11773 in the context of
https://github.com/gardener/gardener/pull/9716.

The problem was that due to a client side optimization, the Gardenlet could no
longer see the ManagedSeed resource of its own seed. This broke the managed seed
detection logic of a monitoring component and hence broke a Prometheus scrape
configuration.

The kubernetes client of the Gardenlet uses a label selector
to filter out the ManagedSeed resources that do not carry the following label:

    name.seed.gardener.cloud/<seed-name>: "true"

https://github.com/gardener/gardener/pull/11965 attempted to fix this issue by
making sure in the mutating webhook that the managed seed resource has this
label. However, this fix was not sufficient, because it did not guarantee that
the managed seed resource is mutated and hence the mutating webhook is called.
There are no other mechanisms that would regularly mutate the managed seed
resource, so even after deploying the fix and waiting a few hours, the label was
missing.

In a development environment, we manually triggered the mutation with

    for i in $(kubectl get managedseeds --no-headers | awk '{print $1}'); do
      echo $i
      kubectl annotate managedseed $i gardener.cloud/operation=reconcile
    done

    for i in $(kubectl get seeds --no-headers | awk '{print $1}'); do
      echo $i
      kubectl annotate seed $i gardener.cloud/operation=reconcile
    done

and this PR makes sure that this mutation is handled automatically by Gardener.

So in a Gardener installation with v1.116 (ok), if v1.117.1 is deployed, the
monitoring component is still going to be broken until the Managed Seed is
mutated and the seed is reconciled some other way. With this PR, these steps are
going to be executed automatically.

This PR adds tasks to the gardener controller manager's startup sequence.

- iterate over all the ManagedSeed resources and
- if the seed name label is missing on a ManagedSeed resource,
  - send an empty patch request. As a side effect, the mutating webhook of
    https://github.com/gardener/gardener/pull/11965 will add the missing label.

  - furthermore, the snippet triggers the reconciliation of the Seed resource.
    - The seed reconciliation flow should be executed to fix the issue with the
      monitoring configuration.
    - Changing only the ManagedSeed resource does not trigger the reconciliation
      of the Seed resource.

  - Patching the ManagedSeed and the Seed resources is done sequentially, and a
    short sleep is added to make sure that the Gardenlet in the seed has already
    observed the change of the ManagedSeed resource by the time it processes the
    reconciliation request of the Seed resource.

This snippet is added to an existing block of code that also deals with seed
labels and that is scheduled to be deleted after Gardener v1.119 was released.
The scheduled removal after v1.119 is just fine for this scenario as well:

- the regression was introduced in v1.117
- this PR will be regularly released maybe with Gardener v1.118 (or v1.119)
  - it will be backported to v1.117.x (and v1.118.x)

So it is fine to remove this code after v1.119.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>