Sort the legend entries by the current value (fixup?)
If the container is restarted frequently, the legend is polluted with
entries for each "actual usage" series, because the aggregation
was removed in https://github.com/gardener/gardener/pull/12057.

Removing the annotation has many benefits: avoids the spike artifacts
and the change in the color shows the container restarts.

By sorting the legend, we can mitigate the problem of polluted
legend entries.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>